feat: 获取一周所有项目的git log 生产周报

+++ src/index.ts
@@ -1,14 +1,14 @@
 #!/usr/bin/env node
 
 import chalk from 'chalk'
-import { exec } from 'child_process'
 import inquirer from 'inquirer'
-import util from 'util'
+import InquirerAutocompletePrompt from 'inquirer-autocomplete-prompt'
+import { $ } from 'zx'
+import { getGitLib } from './gitlab.js'
 import { TypesPrompt } from './types'
 
 const log = console.log
-const execSync = util.promisify(exec) // 配置异步exec
-inquirer.registerPrompt('autocomplete', require('inquirer-autocomplete-prompt'))
+inquirer.registerPrompt('autocomplete', InquirerAutocompletePrompt)
 
 const promptList = [
   {
@@ -23,8 +23,9 @@ const promptList = [
     name: 'git',
     choices: [
       { value: 1, name: '删除本地/远程分支（支持模糊查询）' },
-      { value: 2, name: '删除非当前所有本地分支' },
+      { value: 2, name: '删除非当前的所有本地分支' },
       { value: 3, name: '更新远程/本地分支' },
+      { value: 4, name: '写周报' },
     ],
   },
   {
@@ -34,9 +35,11 @@ const promptList = [
     pageSize: 10,
     when: ({ git }: any) => git === 1,
     source: async (_answers: any, input = '') => {
-      console.log(input === '')
-      const { stdout, stderr } = await execSync('git branch --all')
-      if (stderr) return log(chalk.red(stderr.toString()))
+      const { stdout, stderr } = await $`git branch --all`
+      if (stderr) {
+        log(chalk.red(stderr.toString()))
+        return ['']
+      }
       const list = stdout
         .split('\n')
         .map((e) => e.replace(/(^\s?.*\*\s)|(^\s*)|(\s*$)/g, ''))
@@ -47,51 +50,59 @@ const promptList = [
   },
   {
     type: 'confirm',
+    message: () => `是否确定删除非当前的所有本地分支`,
+    name: 'checkDelLocal',
+    default: false,
+    when: ({ git }: any) => git === 2,
+  },
+  {
+    type: 'confirm',
     message: ({ branch }: any) => `是否确定删除【${branch}分支】(本地和远程)`,
-    name: 'check',
+    name: 'checkDelAll',
     default: false,
     when: ({ branch }: any) => branch,
   },
 ]
 
-inquirer.prompt(promptList).then((props: TypesPrompt.promptProps) => {
-  const { git, branch, check } = props
+inquirer.prompt(promptList).then(async (props: TypesPrompt.promptProps) => {
+  const { git, branch, checkDelLocal, checkDelAll } = props
   if (git === 1) {
-    if (branch && check) {
+    if (branch && checkDelAll) {
       deleteBranch(branch)
     } else {
       log(chalk.red('已取消删除'))
     }
   } else if (git === 2) {
-    deleteLocalBranch()
+    if (checkDelLocal) {
+      deleteLocalBranch()
+    } else {
+      log(chalk.red('已取消删除'))
+    }
   } else if (git === 3) {
     updateBranch()
+  } else if (git === 4) {
+    getGitLib()
   }
 })
 
 /** 删除本地/远程分支 */
-const deleteBranch = (branch: string) => {
-  const cmd = `git branch -d ${branch} && git push origin --delete ${branch}`
-  exec(cmd, (error, stdout, stderr) => {
-    if (error) return log(chalk.red(error.toString()))
-    log(stdout, stderr, chalk.green('删除成功'))
-  })
+const deleteBranch = async (branch: string) => {
+  const { stderr } =
+    await $`git branch -d ${branch} && git push origin --delete ${branch}`
+  if (stderr) return log(chalk.red(stderr.toString()))
+  log(chalk.green('删除成功'))
 }
 
 /* 删除除当前分支之外的所有本地分支 */
-const deleteLocalBranch = () => {
-  const cmd = `git branch | xargs git branch -d`
-  exec(cmd, (error, stdout, stderr) => {
-    if (error) return log(chalk.red(error.toString()))
-    log(stdout, stderr, chalk.green('删除成功'))
-  })
+const deleteLocalBranch = async () => {
+  const { stderr } = await $`git branch | xargs git branch -d`
+  if (stderr) return log(chalk.red(stderr.toString()))
+  log(chalk.green('删除成功'))
 }
 
 /** 更新远程/本地分支 */
-const updateBranch = () => {
-  const cmd = `git remote update origin --prune`
-  exec(cmd, (error, stdout, stderr) => {
-    if (error) return log(chalk.red(error.toString()))
-    log(stdout, stderr, chalk.green('更新成功'))
-  })
+const updateBranch = async () => {
+  const { stderr } = await $`git remote update origin --prune`
+  if (stderr) return log(chalk.red(stderr.toString()))
+  log(chalk.green('更新成功'))
 }
